Rename `Cache.h` to `CacheImage.h`
The only variable left in this file is `imageCache`.
