+++ appOPHD/States/ReportsState.cpp
@@ -1,6 +1,6 @@
 #include "ReportsState.h"
 
-#include "../Cache.h"
+#include "../CacheImage.h"
 #include "../CacheFont.h"
 #include "../Constants/UiConstants.h"
 #include "../StructureManager.h"
