+++ appOPHD/UI/FactoryListBox.cpp
@@ -6,7 +6,7 @@
 #include "../MapObjects/StructureState.h"
 #include "../MapObjects/Structures/Factory.h"
 
-#include "../Cache.h"
+#include "../CacheImage.h"
 #include "../CacheFont.h"
 #include "../Constants/Strings.h"
