+++ appOPHD/UI/MineOperationsWindow.cpp
@@ -2,7 +2,7 @@
 
 #include "TextRender.h"
 
-#include "../Cache.h"
+#include "../CacheImage.h"
 #include "../OreDepositYieldString.h"
 #include "../Resources.h"
 #include "../Constants/Strings.h"
