+++ appOPHD/UI/Reports/MineReport.cpp
@@ -6,7 +6,7 @@
 #include "../../Constants/Strings.h"
 #include "../../Constants/UiConstants.h"
 
-#include "../../Cache.h"
+#include "../../CacheImage.h"
 #include "../../CacheFont.h"
 #include "../../Resources.h"
 #include "../../StructureManager.h"
