+++ appOPHD/appOPHD.vcxproj
@@ -287,7 +287,7 @@
     <ClCompile Include="WindowEventWrapper.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="Cache.h" />
+    <ClInclude Include="CacheImage.h" />
     <ClInclude Include="CacheFont.h" />
     <ClInclude Include="CacheMusic.h" />
     <ClInclude Include="Constants\Strings.h" />
